requirements.txt
fastmcp>=0.9.0
requests>=2.31.0
python-dotenv>=1.0.0
uvicorn>=0.23.0